Fix Node.Add()
Insertion of a node with a path of the type '....X', where the number of
'.' is equal to nodePrefixMaxSize, followed by the insertion of a node
of nodePrefixMaxSize of the same '.' character, will cause the latter
node to not be marked as a value node.  Mark the node appropriately.

This bug causes directory uploads with files that match the pattern
above to fail silently.  The shorter file gets uploaded correctly, but
the manifest cannot see the file when addressed using
/bzz/<node>/<path>.

Signed-off-by: Iliyan Malchev <malchev@gmail.com>